+++ examples/all_examples.rb
@@ -152,7 +152,6 @@ puts JSON.pretty_generate(result)
 
 result = admin.content.approve(first_post_id)
 puts "\nApprove Content\n"
-puts result
 puts JSON.pretty_generate(result)
 
 first_approved = admin.helper.first_live_post
@@ -219,6 +218,39 @@ puts JSON.pretty_generate(result)
 
 result = admin.content_group.delete(new_group)
 puts "\nDelete Group\n"
+puts JSON.pretty_generate(result)
+
+
+# Content Types
+
+result = admin.content_type.list
+puts "\nContent Types\n"
+puts JSON.pretty_generate(result)
+
+first_content_type_id = result[0]['id']
+
+result = admin.content_type.get(first_content_type_id)
+puts "\nFirst Content Type\n"
+puts JSON.pretty_generate(result)
+
+name = 'API Content Type'
+config = { :related_content => false }
+options = { :icon => 'fa-cloud-upload', :primary_editor => 'image' }
+
+result = admin.content_type.add(name,config,options)
+puts "\nNew Content Type\n"
+puts JSON.pretty_generate(result)
+
+new_content_type_id = result['id']
+
+config = { :content_score => false, :related_content => true }
+
+result = admin.content_type.edit(new_content_type_id,config,{})
+puts "\nEdit Content Type\n"
+puts JSON.pretty_generate(result)
+
+result = admin.content_type.delete(new_content_type_id)
+puts "\nDelete Content Type\n"
 puts JSON.pretty_generate(result)